edit the header of the landing page
Put everything on the page into flex-direction:
columns. This will make it easier to push things
around.

Use justify-content: space-around to bring
elements to the desired location in the header

Give the background a navey blue color

give fonts a white color to contrast the blue
background.

Adjust the logo and links size and fonts.

take off the list style type and the text-decor
of the links.

-------------------------------------------------
Edit html

add new class "header-top" and move the logo and
links inside of it. this will help to move them
in CSS.

space things out to help visualize what needs to
be done.