Finish header content
--------------------------------------------------
html:

edit html to add classes to contain elements of
the body of the header.

add a header-body-container that holds everything
in the body of the header.

Adjust location of classes to fit inside of
their resceptive containers.

Add img class with a image of a mechanic.

--------------------------------------------------
css:

Add header body classes to the css.

Add title, subtext, button, and image to css.

Use .header-body-container to move the elements
where they need to be.

Edit button to correct size and location on
webpage.

Edit img size to match the size of the
.header-body container.

--------------------------------------------------
readme:

Add credits to readme for the image.